Added dropdown hiding when clicking outside
Dropdown gets hidden when user clicks outside of it, the input field or the submit button.
Reappears when input is clicked
Submit button is included because if the input is valid, the dropdown
will be hidden via updating the state. If the input is invalid, an error
message will appear instructing the user to choose an option from the
dropdown, so it should remain visible.